LayoutTests/fast/html/HTMLCollection-set-property-expected.txt

Tests setting properties on HTMLCollection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS collection.length is 3
PASS collection.action is document.getElementById('action')
PASS collection[0] is document.getElementById('length')
PASS collection.testInput is document.getElementById('testInput')
PASS collection.foo is undefined.
collection.length = 'about:blank'
collection[0] = 'bar'
PASS collection.length is 3
document.getElementById('length').remove()
PASS collection.length is 2
collection.action = 'about:blank'
PASS collection.action is document.getElementById('action')
document.getElementById('action').remove()
PASS collection.action is undefined.
collection.testInput = 'test'
PASS collection.testInput is document.getElementById('testInput')
document.getElementById('testInput').remove()
PASS collection.testInput is undefined.
collection.foo = 'test'
PASS collection.foo is "test"
PASS collection[0] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
